[kernel][phys] Realize PhysVmo machinery
This change makes good on long-standing plans to have PhysVmo represent
physical memory as-is to be converted directly into a VMO (rather than
with temporary contents to be copied out into a new VMO). This was
previously blocked on being able to convey these regions to the kernel
as regions to wire and avoid when placing PMM arena bookkeeping.

PhysVMO memory in the hand-off is simply identified as being one of
several special types, abstracted behind PhysHandoff::IsPhysVmoType().
As guardrails against this memory not ending up unwired at the end of
hand-off (especially if it does not have an associated description
elsewhere in the hand-off on how to convert it into a VMO), we have the
PMM put this memory into a dedicated page list and then assert that it
is empty in PmmNode::EndHandoff().

Bug: 347766366
Change-Id: Ifb060e0cbb9e6ba1c854786e2788efa1a309bfaf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104559
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>