[fuchsia-async] Change scopes to join when task finishes
This CL changes it so that scopes are joined when the task finishes
rather than when they finish *and* their handles are polled.  Waiting
for their task handles to be polled shouldn't be necessary and might be
confusing to users.

This also changes `Task::cancel` so that it doesn't need to be polled
for it to actually cancel the task; callers can ignore the returned
future if they like.

Change-Id: I2b2d641e10d110f0f07cf14c536e17ddbe6c7294
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108273
Reviewed-by: Tyler Mandry <tmandry@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>