[driver_manager] Update test fixture's StartDriver impl.
CallComponentStart has been moved to the fixture file
and renamed to DriverHostComponentStart.

In the case of a non-colocated driver using dynamic linking,
we want to call the driver host runner Start before returning
from the StartDriver function. Otherwise the driver won't
yet be started (waiting for the driver host to start),
and the returned started TestDriver will not be a valid
pointer.

Bug: 355233670
Change-Id: I41b91027f229ba43a93525e1dd7dbb9f9b122182
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108572
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>