[debugger] Fix crash with unverified line table entries
During the search for the best match for a file+line based breakpoint,
it was possible for the line table to match an address that corresponds
to no code, which produces an invalid Function symbol. The typical case
for this happening is an address that corresponds to some compiler
generated code.

Fixed: https://fxbug.dev/352164877
Fixed: https://fxbug.dev/360821506
Change-Id: I6776031e14dd51aff97308ae9fcc3b87eac0f6f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110912
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>