[dl] Store dep_names_ in SessionModule
This CL will store the decoded Sonames corresponding to DT_NEEDED
offsets when they are decoded from the loaded ELF file. This set of
names will later be used to look up the modules needed to include in a
resolution list.

Since dep_names_ is stored in the class and scoping is not a factor for
as it was for the local variable, the function returns early if the
reserve operation fails; this makes the code a little more readable.

Bug: 338232267
Change-Id: Ideff5f77d00b45c9990e611cc3e3bedc5c323573
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110341
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>