[kernel][vm] Correct page queues reclaim check
Corrects a debug method for testing if a page is in a reclaimable queue
to not be limited to evictable (i.e. pager backed) VMOs. This
restriction is from a time when the reclaimable queues were actually
called the pager backed queues, but when that was changed this method
was not updated.

Change-Id: I790e95f438598e2f7324df8079abdbff93afff85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110233
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>