[kernel][vm] Refactor loan cancellation
Refactors how the the PhysicalPageProvider performs loan cancellation
and tracks down loaned pages. As the PhysicalPageProvider is now the
only caller of GetCowWithReplaceablePage this method is re-designed
into GetCowForLoanedPage with a narrower contract. This narrower
contract allows for combining the stack ownership wait that use to be
in that method and the PhysicalPageProvider into a single wait in the
PhysicalPageProvider, and allows for the functionality of
GetCowForLoanedPage to be logically simpler.

Run-All-Tests: True
Change-Id: I6710bebeb8c14fc61a8943fa93c28c721b4657ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1107993
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>