[netstack3] Multicast Admin FIDL worker
Outline the `MulticastAdminWorker` task to back the
fuchsia.net.multicast.admin FIDL API. Work is sent to the worker via the
`MulticastAdminEventSink`, ensuring all work is serialized (this
eliminates potential races during device removal).

In addition to the new structure, this CL adds the following
functionality:
  * When a device is removed, any routes referencing it in the multicast
    route table are removed.
  * If a new table controller client connects, but we already have an
    active client, reject the incoming request with `ALREADY_IN_USE`.
  * If a new watch request comes in, but we already have a pending
    request, close the client with `HANGING_GET_ERROR`.
  * Multicast routing is enabled while we have an active table
    controller client: enable routing when the client connects, and
    disable it when the client disconnects.

Unfortunately, none of the FIDL integration tests are newly passing:
this is because they all rely on the ability to forward a multicast
packet, which isn't fully supported yet in NS3 core. For now, rely on
unit test coverage.

Bug: 353330225

Change-Id: Ifc533d137bc3a8481a2a04ce6b40bb9784fb2893
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099372
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>