[ffx][repo] Add autopublish option to server start
This adds an option to server start, that takes a package list
file  be watched and automatically publish the packages as they change.

This is the same implementation as `ffx repository publish --watch`,
just conveniently part of the repo server start command.

It is primarily used in-tree, to simplify the implementation of the bash
wrapper script `fx serve` and to keep the publishing log messages
collected in the same file and streams as the repository server log
messages.

Bug: b/360933107
Change-Id: I0dd16619823eef9a3eeabd713e43c6d27b013c44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109337
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>