[wlan] Fix components double remove test flake
One of the tests in the wlan drivers component library was intended to
test that calling Remove twice in a row would correctly handle the case
of another removal being in progress. Unfortunately this test made some
incorrect assumptions about the internal state of the NetworkDevice
object it was removing. The test assumed that the internal state was not
updated until after the removal notification. This is not the case and
this lead to flakiness when the first removal completed before the
second Remove was called.

Instead of blocking the removal notification, block the entire
dispatcher used for removal with a posted task. This properly prevents
the first removal from completing until a completion is signaled.

Verified by adding a sleep between the removals which always triggered
the flake before the fix and passes after the fix.

Fixed: b/338922519
Multiply: wlan-drivers-components-tests
Change-Id: Ief8e000c30595faf787e214ec55586620c10df5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1111818
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Tan <johntan@google.com>