[bazel_sdk] remove dep on fuchsia_sdk_common
The fuchsia_sdk_common repository was needed for us to share the
sanitizer features but this requires external repos to call the
bazel_rules_deps macro in their workspace files. This required a load of
the fuchsia rules which would trigger a download. In order to move away
from requiring loads in WORKSPACE files we needed to break this
dependency.

BUG: b/346416385
Change-Id: I861c281706de0db780fa10cd7fc91e1edba81ab7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112315
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>