[windowed-stats] Introduce a time series glossary.
This change adds a glossary that describes the concepts and components
of the `series` module. These ideas are not trivial and this provides a
much needed reference.

This is more or less a first draft; expect subsequent changes. It is a
living document that should evolve with `windowed-stats`.

Bug: 350797382
Change-Id: Ieaec72981b8c95913d80a9d3c44222e232730443
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108023
Commit-Queue: Sean Olson <seanolson@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>