[audio][drivers] Encapsulate inspect better
This CL pulls the inspect implementation details out of the
main AudioCompositeServer and RingBufferServer classes to a
dedicated class. It also fixes a bug where only the most
recent SetActiveChannels and Start/Stop calls were saved,
instead of the entire history (as was the intention, like
we do for power transitions).

Bug: b/344659863

Change-Id: I211815f5d6cab41a9659e0d71941d73a661cecdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110527
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>