[storage] Limit out-of-space logging
The stress builders run for a long time and trigger out-of-space
issues. The log messages generated by this can end up getting stored in
memfs and, before this CL, would end up OOMing the system.

This changes the out-of-space log messages so that they're only logged
every 60 seconds which should reduce the amount of log messages and we
shouldn't need to see them any more frequently than that.

Bug: 362841925
Change-Id: I53c4f288ee39ca7cd69f1bddf1bf8db6d8b0c3a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110212
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>