[bazel] Generate @fuchsia_idk repository.
Add a new repository to the Bazel workspace used by the
platform build, to expose the platform IDK as an external
repository (currently named @fuchsia_idk).

This still takes a Ninja-generated IDK export directory
as input, but will rewrite the metadata JSON files to
include labels, instead of relative file paths, when
addressing Ninja-generated artifacts (with a few
exceptions, such as a few files in the sysroot).

A future CL will adjust the implementation of the
fuchsia_sdk_repository() rule to accept this external
repository as a proper input.

Change-Id: I1b2e73799a6b9132e9bfe35e9b5c317b4d641fbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110273
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>