[inspect][graph] Fix issue that was leaving TOMBSTONES behind
There were multiple issues surrounding the RAII semantics of the
Vertices and Edges. In particular, dropping either an edge, a source
vertex or a destination vertex was resulting in the edge being partially
dropped.

This change removes the usage of clone_weak which was the source of
problems here. We may want to entirely work on removing that API or
providing better support for it in the inspect library. It's very easy
to misuse in subtle ways.

Fixed: 363073059
Change-Id: Ic83f92bddd24d0d713cc39fe6e77f17f2a1a6d53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112732
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>