[components] Clarify process launcher error for dynamic linker timeouts
When a new process's job is killed while the process is being launched,
it's possible this causes the attempt to retrieve the dynamic linker to
timeout. In this situation, the error is expected, so use an INFO log
instead of ERROR, and make the error message more clear.

This is a sequel to https://fxrev.dev/1090961. If we discover more
errors that are like this we can follow the same pattern.

Fixed: 358166703
Change-Id: I1f88c6e40c3f4916c80eae37f21cec47f54f8dc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113312
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gary Bressler <geb@google.com>