[kernel][rppm] Introduce PowerState
PowerState is a member of a device(cpu) that belongs to a power domain.

The PowerState sole duty is preserving a coherent state of a meaningful
power level.

A power state is represented as a triplet between <Model, Actual Power
State, Desired Power State>. When a model is updated to a functionally
different model, then both actual and desired power level lose meaning.
It is the job of the PowerState object to determine whether a transition
should be dropped or not based on whether it is still relevant for the
current Model.

Additionally a PowerState can request a `Transition` which generates a
`PendingTransition` encapsulating all the information needed to perform
a transition at a later stage.

Bug:
Change-Id: I46442bc3ee23fa54d1d3edd7fe3af292f4f5810c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089296
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>