[netstack3] Move `confirm_reachable` to `IpSocketHandler`
We need to lookup route to confirm neighbor reachability because we
don't have route caching. Moving it to the IP socket layer will make
it closer to where routing decisions are made, and in the future when
we have route caching, it will also be in the `IpSock`.

Fixed: 361817008
Change-Id: I679689307dd1435a8f785cc8e617d1c6975d7c36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109983
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>