[fx] Fix build-dir level default targets
This aligns the fx scripts with how ffx works so build-dir level
targets can be configured, but are still overridden using
`ffx target default set` or the `--target` flag.

This includes the -d flag to `fx`

ffx uses a prioritized list of configuration to resolve the default target.
Once there is a value, the remainder of the list is ignored.

* Specify the target on the ffx command line with --target.
* If configured, use the user level configuration set by ffx target default set.
  This user level configuration is stored relative to $HOME.
* The default configuration compiled into ffx checks two environment variables:
     "$FUCHSIA_DEVICE_ADDR"
     "$FUCHSIA_NODENAME"
    These are used by Fuchsia infra jobs to specify which target should be used
    when running a specific test FUCHSIA_NODENAME is also set to any
    build-directory default target set by `fx set-device`.
* If only one device is discovered

If there is still no target device identified, or if there are multiple devices discovered, an error is returned.

Bug: b/362754361
Change-Id: Ib0b604fccc26a7d2a45ea833a9af4efa5c2159ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110007
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>