[zircon] Remove the flash stress component
The flash stress component was dependent on FVM which has been
superseded by Fxfs as the default storage architecture, and
as far as I know, we're not actively using it. If and when we
need something to stress test the flash, we can ressurect this
or develop something new.

Bug: 357467643
Change-Id: I2ff5f63d90a536705f1f07c2eb42a97ace9eebe1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110034
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Chris Suter <csuter@google.com>