[ufs] Support UTP Task Management Request
This CL defines a UTP Task Magenemt Request Descriptor and creates a
TaskManagementRequestProcessor class to manage the list and handle the
doorbell and completion.
The Task Management Request is not yet issued to the UFS Device and
will be used in the error handler that will be implemented in the SCSI
libray.

Test: fx test ufs-tests qemu_ufs_test
Change-Id: Ib42cfaeca90fad2d75bafba472ff3b85a8915339
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992272
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Jaeyoon Choi <j_yoon.choi@samsung.com>
Reviewed-by: Christopher Anderson <cja@google.com>