[testing] Add a template for c++ test executables
To enable LTO in the platform, we want to be able to disable LTO
during the link step for test executables, in order to save
build time (this requires the use of FatLTO).

This new template declares that an executable is a test
executable, and should receive different build optimizations
than a standard executable receives.

Change-Id: I25d9f45f5208b96345ce1a27cc2487e937ddd946
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1105436
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Paul Kirth <paulkirth@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>