[netstack3] IpDeviceAddr witnesses "non multicast"
Ensure instances of IpDeviceAddr are not multicast addresses. This
allows us to remove a check for multicast in source address selection.

To facilitate this, introduce a new Witness type to the net-types lib:
`NonMulticastAddr`. Note that `UnicastAddr` was not suitable because the
semantics of "unicastness" are complicated for IPv4.

This also increases the "depth" of the nested witnesses held by
`IpDeviceAddr` to 3. Thus this CL adds support to `impl_nested_witness`
in the net-types lib for this additional witness layer.

Fixed: 1112292

Change-Id: Iefbe2136596c1cf6c87979cb48a7c68a1bc3305a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112092
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>