[power][trace] Power element trace events
Added trace to lease request and level trasnsitions in power broker.

On the client side, added trace to the Daemon so that in the future
the overhead from client calling CurrentLevel::Update to the power
broker starting to handle the fidl can be computed.

Trace is enabled in the src/power/bench integration test and routed
to the daemon and the components created in the test.

Bug: 322026959
Test: CQ and manual test
  fx test power-broker-unittests \
    --test-filter=*test_initialize_current_and_broker_status
  fx test -o power-framework-bench-integration-tests \
    --test-filter=*test_topologytestdaemon_toggle -- --repeat 100000
  while the integration test runs, collect trace (a few seconds) with
  ffx trace start --buffer-size 200 \
    --categories '#default,kernel:ipc,power-broker'

Change-Id: I24e7cb7b2923841b8335329cfb5ceb68c85421f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090133
Reviewed-by: Hai Bi <bihai@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Hai Bi <bihai@google.com>