[starnix] Add wake locks to the kernel
This allows the kernel to keep track of active wake locks and/or active
EPOLLWAKEUPs.

An EventPair is signaled whenever a wake lock is active, and cleared
whenever the number of active locks drops to zero. This allows the
Starnix runner to inspect the event to determine whether or not a wake
lock was acquired during container suspension.

Change-Id: I149b208d9e483fdaea14e9e0ee03f60e090c03b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1112632
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>