[bt][gatt] Remove unnecessary handle tracking
Discoveriing a handle for the service changed characteristic is used to
decide whether to attempt to send indications, but is never cleared when
there are no subscribers, so will likely be on after the systems runs
for any reasonable length of time. Instead continue to use the set of
subscribed peers directly.

Test: code removal only; fx test bt-host-gatt-tests
Change-Id: I6c63ae552df18e182effe5b20a4abe65c90f3f2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009175
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>