[fidl][rust] Slightly improve initialisation of thread-local buffers
We don't need to check if the buffer capacity is zero each time we use
the buffer.  We can just reserve the default capacity in the TlsBuf
constructor, which is run lazily anyway (by
std::thread::LocalKey::with()).  The thread-local buffer's capacity
should never get shrunk.

This will avoid one conditional and reduce the size of this inlined
code a little.

Change-Id: I248684b388890d0ee35b282a0b7de394f4de9e30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113476
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>