[fuchsia_fs] Indicate open1 functions are deprecated (1/2)
Add open3 version of functions that use open. Also, change the names of
functions that use open1 in fuchsia_fs to include `deprecated`.

The critical changes are in src/lib/fuchsia-fs.

Next steps for transitioning fuchsia_fs to use open3 is to migrate all
callers of the deprecated open1 functions to use open3, and eventually
removing the deprecating functions.

This change is split into two as there are out-of-tree callers of some
of these functions. Namely, `directory::open_file`,
`directory::open_file_no_describe`, and `directory::open_in_namespace`.
For these three functions, we need to migrate the out-of-tree callsites
to call `open_X_deprecated` before updating the `open_X` functions to
use open3 in the second part of this change.

Update tests in fuchsia-fs to call functions that are using open3
instead of the deprecated open1.

Bug: 361450366
Test: fx test fuchsia-fs-tests
Change-Id: I2bf4ea98e7e012ca1f89a148bee8922aa07f46d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1109012
Owners-Override: Adam Barth <abarth@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>