[netstack3] Remove TCP Takeable trait
The takeable trait defers to implementers a side-effect of how the TCP
state machine is implemented: moving references around when matching
enums is hard to do while keeping the borrow checker happy.

The existing implementation relies on `Default` impls to leave behind
supposedly unused memory, but there's no real checking that the memory
is not really used, which can cause bugs and hurts evolution. It's
better to have strong type safe statements of everywhere we need to take
things, even if it costs an option check at runtime.

While I'm here, also remove the clone bounds from the buffer providers,
it was doing mostly the same thing as Takeable but in less obvious ways.

Also fixed a bug where `on_icmp_error` would not consider the `Closing`
state. Unfortunately there's no good test coverage for this, but it's
fixed thanks to type steering. Test coverage will be added in a future
change.

Note that the attached bug suggested using `replace-with` instead of
Takeable. `replace-with` has 2 problems that called for a different
solution:

a) Ergonomics for _optionally_ replacing (like in the case of errors)
are very poor.
b) `replace-with` calls for copying the structure out and back in even
in the case where it doesn't get replace, which is the common case when
handling segments, for example. The data copies of these large structs
would've been a bigger performance hit than the Option check within the
new Takeable type.

Fixed: 42082051
Change-Id: I4780513b7d720363a16e2248714120d73e4ed0b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113413
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>