[starnix] Add wake leases to runner
This change updates the starnix runner to acquire wake leases for all
the containers it runs. These are dropped when the container is
suspended, and reacquired before the container is resumed.

The change is hidden behind a feature that is currently disabled.

Change-Id: I31fa2adbae1c83a58295e99a5188a7db7ee73d77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113914
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>