[build][api] Fix `print --last-build-only product_bundles`.
The command was always returning an empty list, even when building
an actual product bundle in the last build invocation, because
it filtered based on the "path" value of each entry, which points
to a directory that never appears as an output path in the Ninja
build plan.

To fix this, filter on the "json" field instead, which points to
the actual product bundle JSON manifest file, a real Ninja artifact.

Bug: 365039385
Change-Id: Ic490839b90bf514672e6449d66a224905dc6e421
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114412
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>