[wlan][policy] record if network is single BSS
This change records whether multiple BSS's have been seen for a network
so that roaming can avoid scanning when there are no other BSS's to
choose from. The implementation will wait for a few scans that have only
one BSS before deciding that the network only has one BSS, and it only
pays attention to the network that is being scanned for. The data is not
persisted.

Test: unit tests: fx test wlancfg-tests
Change-Id: I92f15c5cdbdb8a15526adc90e0e0a938dbb8fc2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110344
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>