[perf_publish] Add simpler interface for using with data packages
Add a simpler interface for using publish_fuchsiaperf() when the
metric name expectation file is provided by a data package.  The
caller passes a reference to the data package module instead of
getting the file out of the module themselves.

Convert one of the tests to use this interface.

This fixes the case of using FUCHSIA_EXPECTED_METRIC_NAMES_DEST_DIR to
update expectation files.  I tested that by running the following and
checking the timestamp of the expectation file:

  fx test -o --e2e perf_publish_example_test.sh \
    --env FUCHSIA_EXPECTED_METRIC_NAMES_DEST_DIR=$(pwd)/src/tests/end_to_end/perf/expected_metric_names/

Bug: 358492600
Bug: 333140597
Change-Id: Ibb30557092a950a46d8bd54bccba1afd1bc7345e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114121
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Kevin Cho <chok@google.com>