[paver] expose boot attempts
Expose a new paver FIDL API that returns the slot status as well as the
number of times it has attempted boot.

Eventually we may want to deprecate the API that only returns slot
status and use this everywhere, but that can be a follow-on effort if we
decide it's worth the time.

Bug: b/349617985
Change-Id: I13bee0a5bf4ab4ee9023f35c9aa8339394335ce4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110009
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
API-Review: Chris Suter <csuter@google.com>