[driver_manager] Implement DynamicLinkerDriverHostComponent Start.
This makes a call to fuchsia_driver_loader::LoadDriver,
(currently returns ZX_ERR_NOT_SUPPORTED).

Adds a partial fake driver host, to be completed in a future CL.
This is so the bootstrap channel is not dropped prematurely,
leading to the node beginning removal and the LoadDriver FIDL
request not completing.

Bug: 341998660

Change-Id: I67beef104bb536a24c45b7718ac51ba6f423dcbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093155
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>