[power] Add a gni file for node_config template
Move node_config gn template to node_config.gni file. This allows the
template to be used in integration tests.
Add an output arg to the template. This allows the template to be used
in unit tests to validate the node config files.

This is required by tqr/880123

Bug: b/362554191
Change-Id: Ib49c803cc6819b62cbe4f9147b578030ab98a56b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1110653
Reviewed-by: Hai Bi <bihai@google.com>
Commit-Queue: Didi She <didis@google.com>