[gigaboot] on QEMU, support on-disk commandline
We need a way to tell Gigaboot to stop in fastboot, and using the
standard UEFI commandline mechanism has proven difficult in our emulator
setup. This CL adds support for QEMU only to look for commandline
contents in a partition named "qemu-commandline", so that tests can
create and populate this partition if needed to pass information into
Gigaboot.

Bug: b/359557990
Change-Id: Ic50d78d667a86c1fbf59626dfa2d744f6afeb611
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114653
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Dov Shlachter <dovs@google.com>