[fasttime] Use sequential read for monotonic ticks
The mono_ticks_offset in the vDSO is no longer constant, as it will be
updated during periods of suspension. Therefore, it must now be read
with a sequential read just like the kernel version of the same routine.

Note that the register read ordering is still not correctly enforced;
this will need to be fixed in future CLs.

Bug: 341785588
Change-Id: I68303e6d0cc8ad868e0019ff1d0d11e804d73028
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114705
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: John Grossman <johngro@google.com>