[vsock] Remove SendVmo functionality
This removes the ability to sent entire VMOs through a virtio socket in
the both the service and driver layer. This functionality is currently
unused and untested - the data path used by sockets is a Zircon socket.

In an earlier version of the system this data path was used to more
efficiently send large amounts of data from a file into a socket. If and
when this use case arises again we can either revive this logic or find
a new way to optimize the stack as many other parts of the system behave
differently from when this analysis was last performed.

Change-Id: Ic18519e15a836f448192b65ca27c8d31aad432c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114036
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Tim Kilbourn <tkilbourn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>