[build][python] Allow empty data packages
Produce a data package if data_package_name is set, even if
data_sources is empty.

This is useful if we have multiple test cases that share a Python
module that does "import test_data", and some of the test cases have
empty data packages.  This can arise if some of the test cases have
perf metric expectation files and some don't, which happens in
tqr/881658.

Bug: 358492600
Bug: 333140597
Change-Id: Ifdbfa429eb721991df6e31288cedc787f013da3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114817
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>