[device-enumeration] Don't verify partitions
Which partitions get bound is more of a question of whether fshost is
doing its job properly, and we have extensive tests for fshost, so
remove these from device-enumeration.

This is necessary for enabling storage-host, since that feature will be
enabled on a per-product basis, not a per-board basis (and so we'd have
to change which devices we look for depending on which product is
running the test, which is tricky because the test is compiled the same
regardless of the product it is being compiled for).

Bug: 339491886
Change-Id: I0b9f306f48450e9a8c393e5b84af196ad7364734
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115814
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Garratt Gallagher <garratt@google.com>