[bazel] define sysroot filegroups in fuchsia_sdk
This CL defines the sysroot filegroups in the fuchsias_sdk repository
instead of the fuchsia_clang repository. This allows us to remove the
symlink from the fuchsia_clang repository.

If we keep a symlink in the fuchsia_clang repository then that symlink
will materialize when we ship the fuchsia_clang repository. This would
put the sysroot in the clang repository where it does not belong.

BUG: b/346416385
Change-Id: I020d0f759cc46190f49f1d0187c8e90febc13b74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1113545
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>