[cobalt][fidl] Delete removed FIDL elements
This FIDL library is not in the SDK, so it only needs to support HEAD.
This also correctly marking the entire library `@available(added=HEAD)`.

`@available(added=7)` was added in https://fxrev.dev/664072 and not
removed when the category was changed in https://fxrev.dev/792622.

Fixed: 315496857
Bug: 364628320
Change-Id: If0ad0cb20f25dab9cd4583ae857e68c298c51d46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1114041
Commit-Queue: David Dorwin <ddorwin@google.com>
API-Review: Francois Rousseau <frousseau@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>