[recovery] Add recovery to Assembly
This allows Assembly to be able to include the recovery-fdr components
in the assembled image.

As part of that, it removes a number of unused GN args that changed
the contents of the recovery packages.  The next step in this process
is to move configuration to a domain config package and config-
capabilities.

Bug: b/362365879
Change-Id: I4c89326f89e360f0131327a426057822bbf0d23b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1108586
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>