[debugger] Treat non-symbol index servers as public
With the introduction of DEBUGINFOD_URLS support in symbolizer we should
stop assuming that servers coming from anywhere other than the symbol
index have authentication.

If we have a need to specify servers that require authentication on the
command line then we can add a new argument for those. Debuginfod was
never intended to be used with non-public servers and so these will
always be assumed to be public.

This also fixes a crash in zxdb when trying to initialize a
std::filesystem::path with a null pointer (specifically when "HOME" is
not set in the environment).

Bug: 365193885
Change-Id: Iba114d510bcc70053afef443bd0d883c23bf0a8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1115593
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>