[bazel][assembly] Add missing "json" metadata field for product bundles.
The schema of the `product_bundles` build API module states that
the "json" attribute, which points to the product bundle manifest
file, is required.

However, the bazel_product_bundle() didn't define this in the
metadata it generated for Bazel-assembled product bundles. This
caused `build/api/client print --last-build-only product_bundles`
to fail, because it could not find the "json" object field when
trying to filter the content of `out/default/product_bundles.json`.

See the CQ failures on [1] for examples.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1114414/2
Bug: 42175326

Change-Id: Ie3bfc5fd1f72201dc68ae4c6ada0765c7da4fd06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116274
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>